requirements.txt
google-generativeai
streamlit
python-dotenv
PyPDF2
pdf2image
python-docx
Pillow
poppler-utils